moved everything in ChatBox.vue onMount() to Chat.vue
